Hoist inner classes out of `DocGenerator`.
The tool has its own package now, so this nesting is not valuable.
