+++ src/main/java/dev/ionfusion/fusion/_private/doc/tool/DocGenerator.java
@@ -13,19 +13,11 @@ import dev.ionfusion.fusion.FusionException;
 import dev.ionfusion.fusion.FusionRuntime;
 import dev.ionfusion.fusion.ModuleIdentity;
 import dev.ionfusion.fusion._private.HtmlWriter;
-import dev.ionfusion.fusion._private.doc.model.BindingDoc;
 import dev.ionfusion.fusion._private.doc.model.DocTreeNode;
-import dev.ionfusion.fusion._private.doc.model.ModuleDocs;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.Arrays;
-import java.util.Comparator;
-import java.util.Map;
-import java.util.Map.Entry;
-import java.util.Set;
-import java.util.TreeSet;
 import java.util.function.Predicate;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -36,7 +28,7 @@ import java.util.regex.Pattern;
 public final class DocGenerator
 {
     /** HTML content for the masthead links */
-    private static final String HEADER_LINKS =
+    static final String HEADER_LINKS =
         "<div class='indexlink'>" +
         "<a href='index.html'>Top</a> " +
         "<a href='binding-index.html'>Binding Index</a> " +
@@ -214,549 +206,10 @@ public final class DocGenerator
     }
 
 
-    //========================================================================
-
-
-    private static final class ModuleWriter
-        extends HtmlWriter
-    {
-        private final Predicate<ModuleIdentity> myFilter;
-        private final String                    myBaseUrl;
-        private final DocTreeNode               myDocTreeNode;
-        private final ModuleDocs                myModuleDocs;
-        private final ModuleIdentity            myModuleId;
-        private final MarkdownProcessor         myMarkdown = new MarkdownProcessor();
-
-        public ModuleWriter(Predicate<ModuleIdentity> filter,
-                            File      outputFile,
-                            String    baseUrl,
-                            DocTreeNode doc)
-            throws IOException
-        {
-            super(outputFile);
-            myFilter  = filter;
-            myBaseUrl = baseUrl;
-            myDocTreeNode = doc;
-            myModuleDocs = doc.getModuleDocs();
-            myModuleId = myModuleDocs.getIdentity();
-        }
-
-        void renderModule()
-            throws IOException
-        {
-            openHtml();
-            {
-                String modulePath = myModuleId.absolutePath();
-                renderHead(modulePath, myBaseUrl, "common.css", "module.css");
-
-                openBody();
-                {
-                    renderHeader();
-                    renderModuleIntro();
-                    renderSubmoduleLinks();
-                    renderBindings();
-                }
-                closeBody();
-            }
-            closeHtml();
-        }
-
-
-        private void renderModulePathWithLinks(ModuleIdentity id)
-            throws IOException
-        {
-            ModuleIdentity parent = id.parent();
-            if (parent != null)
-            {
-                renderModulePathWithLinks(parent);
-            }
-
-            append('/');
-
-            String baseName = id.baseName();
-
-            if (id == myModuleId)
-            {
-                // Don't link to ourselves, that's silly.
-                append(baseName);
-            }
-            else
-            {
-                linkToModule(id, baseName);
-            }
-        }
-
-        private void renderHeader()
-            throws IOException
-        {
-            append(HEADER_LINKS);
-            append("<h1>Module ");
-            renderModulePathWithLinks(myModuleId);
-            append("</h1>");
-        }
-
-        private void renderModuleIntro()
-            throws IOException
-        {
-            String overview = myModuleDocs.getOverview();
-            if (overview != null)
-            {
-                markdown(overview);
-            }
-        }
-
-        private void renderSubmoduleLinks()
-            throws IOException
-        {
-            Map<String, DocTreeNode> submodules = myDocTreeNode.submoduleMap();
-            if (submodules == null || submodules.isEmpty()) return;
-
-            renderHeader2("Submodules");
-
-            String[] names = submodules.keySet().toArray(new String[0]);
-            Arrays.sort(names);
-
-            append("<ul class='submodules'>");
-            for (String name : names)
-            {
-                ModuleDocs module = submodules.get(name).getModuleDocs();
-
-                String escapedName = escapeString(name);
-                append("<li>");
-                linkToModule(module.getIdentity(), escapedName);
-
-                String oneLiner = module.getOneLiner();
-                if (oneLiner != null)
-                {
-                    append(" &ndash; <span class='oneliner'>");
-                    markdown(oneLiner);
-                    append("</span>");
-                }
-                append("</li>\n");
-            }
-            append("</ul>\n");
-        }
-
-
-        private void renderBindingIndex(String[] names)
-            throws IOException
-        {
-            if (names.length == 0) return;
-
-            append("<div class='exports'>\n");
-            for (String name : names)
-            {
-                String escapedName = escapeString(name);
-                linkToBindingAsName(myModuleId, escapedName);
-                append("&nbsp;&nbsp;\n");
-            }
-            append("</div>\n");
-        }
-
-
-        private void renderBindings()
-            throws IOException
-        {
-            Map<String, BindingDoc> bindings = myModuleDocs.getBindingDocs();
-            if (bindings == null || bindings.isEmpty()) return;
-
-            renderHeader2("Exported Bindings");
-
-            String[] names = bindings.keySet().toArray(new String[0]);
-            Arrays.sort(names, new BindingComparator());
-
-            renderBindingIndex(names);
-
-            for (String name : names)
-            {
-                // May be null:
-                BindingDoc binding = bindings.get(name);
-                renderBinding(name, binding);
-            }
-        }
-
-
-        /* CSS hierarchy:
-         *
-         *  binding
-         *    name
-         *    kind
-         *    doc
-         *      oneliner -- presently unused, intended to be text description
-         *      body
-         *      also
-         */
-        private void renderBinding(String name, BindingDoc doc)
-            throws IOException
-        {
-            String escapedName = escapeString(name);
-
-            append("\n<div class='binding' id='");
-            append(escapedName);
-            append("'><span class='name'>");
-            append(escapedName);
-            // FIXME The </a> below is spurious, but unit tests don't flag it.
-            append("</a></span>");   // binding div is still open
-
-            if (doc == null)
-            {
-                append("<p class='nodoc'>No documentation available.</p>\n\n");
-            }
-            else
-            {
-                if (doc.getKind() != null)
-                {
-                    append(" <span class='kind'>");
-                    // Using enum toString() allows display name to be changed
-                    append(doc.getKind().toString().toLowerCase());
-                    append("</span>\n");
-                }
-
-                append("<div class='doc'>");
-
-                if (doc.getUsage() != null || doc.getBody() != null)
-                {
-                    StringBuilder buf = new StringBuilder();
-
-                    if (doc.getUsage() != null)
-                    {
-                        buf.append("    ");
-                        buf.append(doc.getUsage());
-                        buf.append('\n');
-                    }
-
-                    if (doc.getBody() != null)
-                    {
-                        buf.append('\n');
-                        buf.append(doc.getBody());
-                        buf.append('\n');
-                    }
-
-                    append("<div class='body'>");
-                    markdown(buf.toString());
-                    append("</div>\n");
-                }
-
-                append('\n');
-
-
-                ModuleIdentity[] ids =
-                    doc.getProvidingModules().toArray(new ModuleIdentity[0]);
-                Arrays.sort(ids);
-
-                boolean printedOne = false;
-                for (ModuleIdentity id : ids)
-                {
-                    if (id != myModuleId && myFilter.test(id))
-                    {
-                        if (printedOne)
-                        {
-                            append(", ");
-                        }
-                        else
-                        {
-                            append("<p class='also'>Also provided by ");
-                        }
-
-                        linkToBindingAsModulePath(id, escapedName);
-                        printedOne = true;
-                    }
-                }
-                if (printedOne)
-                {
-                    append("</p>\n");
-                }
-
-                append("</div>\n"); // doc
-            }
-
-            append("</div>\n"); // binding
-        }
-
-
-        private void markdown(String text)
-            throws IOException
-        {
-            String md = myMarkdown.markdown(text);
-            append(md);
-        }
-    }
-
-
-    //========================================================================
-
-
-    private static final class IndexWriter
-        extends HtmlWriter
-    {
-        private final Predicate<ModuleIdentity> myFilter;
-
-
-        public IndexWriter(Predicate<ModuleIdentity> filter, File outputFile)
-            throws IOException
-        {
-            super(outputFile);
-            myFilter = filter;
-        }
-
-
-        void renderIndex(DocIndex index)
-            throws IOException
-        {
-            openHtml();
-            {
-                renderHead("Fusion Binding Index", null, "common.css", "index.css");
-
-                openBody();
-                {
-                    append("<div class='indexlink'>" +
-                           "<a href='index.html'>Top</a> " +
-                           "<a href='permuted-index.html'>Permuted Index</a>" +
-                           "</div>\n");
-
-                    renderHeader1("Binding Index");
-
-                    append("<table><tbody>");
-                    for (Entry<String, Set<ModuleIdentity>> entry
-                        : index.getNameMap().entrySet())
-                    {
-                        String escapedName = escapeString(entry.getKey());
-                        append("<tr><td class='bound'>");
-                        append(escapedName);
-                        append("</td><td>");
-
-                        boolean printedOne = false;
-                        for (ModuleIdentity id : entry.getValue())
-                        {
-                            if (myFilter.test(id))
-                            {
-                                if (printedOne)
-                                {
-                                    append(", ");
-                                }
-                                linkToBindingAsModulePath(id, escapedName);
-                                printedOne = true;
-                            }
-                        }
-
-                        append("</td></tr>\n");
-                    }
-                    append("</tbody></table>\n");
-                }
-                closeBody();
-            }
-            closeHtml();
-        }
-    }
-
-
-    //========================================================================
-
-
-    private static final class PermutedIndexWriter
-        extends HtmlWriter
-    {
-        private final Predicate<ModuleIdentity> myFilter;
-        private final DocIndex                  myIndex;
-
-        /** Maps keywords to the lines in which they exist. */
-        private final TreeSet<Line> myLines;
-
-
-        /** An index line. */
-        private static final class Line
-            implements Comparable<Line>
-        {
-            private final String myPrefix;
-            private final String myKeyword;
-            private final Entry<String, Set<ModuleIdentity>> myEntry;
-
-
-            Line(Entry<String, Set<ModuleIdentity>> entry,
-                 int keywordStartPos,
-                 int keywordLimitPos)
-            {
-                String name = entry.getKey();
-                myPrefix  = name.substring(0, keywordStartPos);
-                myKeyword = name.substring(keywordStartPos, keywordLimitPos);
-                myEntry   = entry;
-            }
-
-            String bindingName()
-            {
-                return myEntry.getKey();
-            }
-
-            String prefix()
-            {
-                return myPrefix;
-            }
-
-            String keyword()
-            {
-                return myKeyword;
-            }
-
-            String suffix()
-            {
-                int pos = myPrefix.length() + myKeyword.length();
-                return bindingName().substring(pos);
-            }
-
-            Set<ModuleIdentity> modules()
-            {
-                return myEntry.getValue();
-            }
-
-            @Override
-            public int compareTo(Line that)
-            {
-                int result = myKeyword.compareTo(that.myKeyword);
-                if (result == 0)
-                {
-                    result = myPrefix.compareTo(that.myPrefix);
-                    if (result == 0)
-                    {
-                        // We shouldn't get this far often, so we spend time to
-                        // get the suffix rather that memory to cache it.
-                        result = suffix().compareTo(that.suffix());
-                    }
-                }
-                return result;
-            }
-        }
-
-
-        public PermutedIndexWriter(Predicate<ModuleIdentity> filter,
-                                   DocIndex index,
-                                   File outputFile)
-            throws IOException
-        {
-            super(outputFile);
-
-            myFilter = filter;
-            myIndex  = index;
-            myLines  = new TreeSet<>();
-        }
-
-
-        private void permute()
-        {
-            for (Entry<String, Set<ModuleIdentity>> entry
-                     : myIndex.getNameMap().entrySet())
-            {
-                String name = entry.getKey();
-
-                int pos = 0;
-                while (true)
-                {
-                    int underscorePos = name.indexOf('_', pos);
-                    if (underscorePos == -1)
-                    {
-                        myLines.add(new Line(entry, pos, name.length()));
-                        break;
-                    }
-                    else if (pos < underscorePos)
-                    {
-                        myLines.add(new Line(entry, pos, underscorePos));
-                    }
-                    pos = underscorePos + 1;
-                }
-            }
-        }
-
-
-        void renderIndex()
-            throws IOException
-        {
-            permute();
-
-            openHtml();
-            {
-                renderHead("Fusion Binding Index (Permuted)", null, "common.css", "index.css");
-
-                openBody();
-                {
-                    append("<div class='indexlink'>" +
-                           "<a href='index.html'>Top</a> " +
-                           "<a href='binding-index.html'>Alphabetical Index</a>" +
-                           "</div>\n");
-
-                    renderHeader1("Permuted Binding Index");
-
-                    append("<table><tbody>");
-                    for (Line line : myLines)
-                    {
-                        String escapedName = escapeString(line.bindingName());
-
-                        append("<tr><td class='prefix'>");
-                        escape(line.prefix());
-                        append("</td><td class='tail'><span class='keyword'>");
-                        escape(line.keyword());
-                        append("</span>");
-                        escape(line.suffix());
-                        append("</td><td>");
-
-                        boolean printedOne = false;
-                        for (ModuleIdentity id : line.modules())
-                        {
-                            if (myFilter.test(id))
-                            {
-                                if (printedOne)
-                                {
-                                    append(", ");
-                                }
-                                linkToBindingAsModulePath(id, escapedName);
-                                printedOne = true;
-                            }
-                        }
-
-                        append("</td></tr>\n");
-                    }
-                    append("</tbody></table>\n");
-                }
-                closeBody();
-            }
-            closeHtml();
-        }
-    }
-
-
     private static void log(String message)
     {
         System.out.print(Timestamp.now());
         System.out.print(" ");
         System.out.println(message);
-    }
-
-
-    //========================================================================
-
-
-    /**
-     * Customer comparator to hide ugly #% bindings down at the bottom of
-     * binding lists. Otherwise, they tend to show up early, which is silly
-     * since most people don't care about them and shouldn't use them.
-     */
-    private static final class BindingComparator
-        implements Comparator<String>
-    {
-        @Override
-        public int compare(String arg0, String arg1)
-        {
-            if (arg0.startsWith("#%"))
-            {
-                if (! arg1.startsWith("#%"))
-                {
-                    return 1;
-                }
-            }
-            else if (arg1.startsWith("#%"))
-            {
-                return -1;
-            }
-
-            return arg0.compareTo(arg1);
-        }
     }
 }
